Akairo conversion
