+++ data.example.json
@@ -4,7 +4,7 @@
     "id": "",
     "secret": ""
   },
-  "botstatToken": "<Token>",
+  "botstatToken": "<Token>", // Closed API. Request key or leave empty
 	"prefix": ">",
 	"owners": ""
 }
