--- src/Bot.ts
@@ -1,5 +1,0 @@
-import configFile from './config'
-import BotClient from './client/BotClient'
-
-const client = new BotClient({ token: configFile.clientToken, owners: configFile.owners })
-client.start()
