+++ src/client/BotClient.ts
@@ -1,8 +1,9 @@
 /* eslint-disable no-unused-vars */
 import { AkairoClient, CommandHandler, ListenerHandler } from 'discord-akairo'
-import { User, Message } from 'discord.js'
+import { User, Message, ActivityType, ActivityOptions } from 'discord.js'
 import * as path from 'path'
 import axios, { AxiosInstance } from 'axios'
+import { WebhookLogger } from '../structures/WebhookLogger'
 import configFile from '../config'
 
 declare module 'discord-akairo' {
@@ -19,7 +20,9 @@ interface BotOptions {
 
 export default class BotClient extends AkairoClient {
   public config: BotOptions;
-  public botstat: AxiosInstance
+  public botstat?: AxiosInstance;
+  public logger: WebhookLogger;
+
   public listenerHandler: ListenerHandler = new ListenerHandler(this, {
     directory: path.join(__dirname, '..', 'events')
   })
@@ -52,11 +55,15 @@ export default class BotClient extends AkairoClient {
     })
 
     this.config = config
-    this.botstat = axios.create({
-      baseURL: 'https://tmuniversal-api.herokuapp.com/api/v1',
-      timeout: 5000,
-      headers: { Authorization: `Bearer ${configFile.botstatToken}` }
-    })
+    this.logger = WebhookLogger.instance
+
+    if (configFile.botstatToken && configFile.botstatToken?.length !== 0) {
+      this.botstat = axios.create({
+        baseURL: 'https://tmuniversal-api.herokuapp.com/api/v1',
+        timeout: 5000,
+        headers: { Authorization: `Bearer ${configFile.botstatToken}` }
+      })
+    } else this.botstat = undefined
   }
 
   private async _init (): Promise<void> {
@@ -71,10 +78,74 @@ export default class BotClient extends AkairoClient {
     this.listenerHandler.loadAll()
   }
 
-  public async start (): Promise<string> {
+  public async start (): Promise<BotClient> {
     // eslint-disable-next-line no-console
-    console.log('Starting')
+    console.log('Starting the bot...')
     await this._init()
-    return this.login(this.config.token)
+    await this.login(this.config.token)
+
+    const users = this.users.cache.size
+    const channels = this.channels.cache.size
+    const guilds = this.guilds.cache.size
+
+    await this.updateBotStats(guilds, channels, users)
+
+    this.setInterval(() => this.changeStatus(), 120000)
+
+    // Error handling
+    this.on('error', e => this.logger.error(e.message))
+    this.on('warn', w => this.logger.warn(w))
+
+    //  Process handling
+    process.once('SIGINT', () => {
+      this.logger.info(`[${this.user.username}] Process exit.`)
+      this.destroy()
+    })
+    process.on('exit', () => {
+      this.logger.info(`[${this.user.username}] Process exit.`)
+      this.destroy()
+    })
+    process.on('uncaughtException', (err: Error) => {
+      const errorMsg = (err ? err.stack || err : '').toString().replace(new RegExp(`${__dirname}/`, 'g'), './')
+      this.logger.error(errorMsg)
+    })
+    process.on('unhandledRejection', (err: Error) => {
+      this.logger.error('Uncaught Promise error: \n' + err.stack)
+    })
+
+    return this
+  }
+
+  public async changeStatus (options?: ActivityOptions) {
+    const users = this.users.cache.size
+    const channels = this.channels.cache.size
+    const guilds = this.guilds.cache.size
+
+    const statuses: Array<ActivityOptions> = [
+      { type: 'PLAYING', name: `with ${users} users` },
+      { type: 'LISTENING', name: `${users} users` },
+      { type: 'WATCHING', name: `over ${users} users` },
+      { type: 'PLAYING', name: `in ${guilds} servers` },
+      { type: 'WATCHING', name: 'tmuniversal.eu' },
+      { type: 'PLAYING', name: `${configFile.prefix}help for help` },
+      { type: 'WATCHING', name: `${guilds} servers` }
+    ]
+
+    const chooseStatus = options || statuses[~~(Math.random() * statuses.length)]
+    const details: ActivityOptions = { type: chooseStatus.type || 'PLAYING' as ActivityType }
+    if (chooseStatus.url) details.url = chooseStatus.url
+    this.user.setActivity(chooseStatus.name, details)
+  }
+
+  public async updateBotStats (guilds: number, channels: number, users: number) {
+    if (!this.botstat) return Promise.resolve(this.logger.warn('Botstat API is disabled'))
+    return this.botstat.patch('/botstat/' + this.user.id, {
+      guilds: guilds,
+      channels: channels,
+      users: users
+    })
+      // eslint-disable-next-line no-console
+      .then(() => console.info('Uploaded user base stats to API.'))
+      .catch(e => this.logger.error(e))
   }
 }
